[improvement] DatetimePicker: optimize render (#3996)
